refactor(unity-project): split editor path responsibilities and share path exception helper
Why:
- UnityEditorPathResolver に探索/検証/正規化が集中しており、内部凝集が下がっていたため。
- パス例外判定が複数クラスへ重複しており、責務が散在していたため。

What Changed:
- UnityEditorPathResolver をオーケストレーター化し、実行ファイル探索を UnityEditorExecutablePathLocator、バージョン整合検証を UnityEditorVersionConsistencyValidator へ分離した。
- PathFormatExceptionHelper を追加し、UnityProjectResolver と UnityVersionResolver と locator で共通利用するようにした。
- 既存の IUnityEditorPathResolver 契約と戻り値契約は維持した。

Impact:
- 変更理由ごとの責務境界が明確になり、batchmode 連携時の拡張点が限定される。
- パス例外分類ロジックの一元化で将来の修正漏れリスクを抑制する。
